server/requirements.txt

Flask-Session
Flask
Flask-Dance
flask-cors
requests-oauthlib
PyJWT
cryptography
